% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{accept_family_binary}
\alias{accept_family_binary}
\title{Compute the acceptable family for binary data}
\usage{
accept_family_binary(
  post_y_pred,
  post_lpd,
  XX,
  indicators,
  eps_level = 0.05,
  eta_level = 0,
  loss_type = "cross-ent",
  yy = NULL,
  post_y_hat = NULL,
  K = 10,
  sir_frac = 0.5,
  plot = TRUE
)
}
\arguments{
\item{post_y_pred}{\code{S x n} matrix of posterior predictive
at the given \code{XX} covariate values}

\item{post_lpd}{\code{S} evaluations of the log-likelihood computed
at each posterior draw of the parameters}

\item{XX}{\code{n x p} matrix of covariates at which to evaluate}

\item{indicators}{\code{L x p} matrix of inclusion indicators
where each row denotes a candidate subset}

\item{eps_level}{probability required to match the predictive
performance of the "best" model (up to \code{eta_level})}

\item{eta_level}{allowable margin (%) between each acceptable model
and the "best" model}

\item{loss_type}{loss function to be used; currently consider
"cross-ent" (cross-entropy) or "misclass" (misclassication rate)}

\item{yy}{\code{n}-dimensional vector of response variables (optional)}

\item{post_y_hat}{\code{S x n} matrix of posterior fitted values
at the given \code{XX} covariate values (optional)}

\item{K}{number of cross-validation folds}

\item{sir_frac}{fraction of the posterior samples to use for SIR}

\item{plot}{logical; if TRUE, include a plot to summarize the predictive
performance across candidate subsets}
}
\value{
a list containing the following elements:
\itemize{
\item \code{all_accept}: indices (i.e., rows of \code{indicators})
that correspond to the acceptable subsets
\item \code{beta_hat_small} linear coefficients for the
smallest acceptable model
\item \code{beta_hat_min} linear coefficients for the
"best" acceptable model
\item \code{ell_small}: index (i.e., row of \code{indicators}) of the
smallest acceptable model
\item \code{ell_min}: index (i.e., row of \code{indicators}) of the
"best" acceptable model
}
}
\description{
Given output from a Bayesian model and a candidate of
subsets, compute the *acceptable family* of subsets that
match or nearly match the predictive accuracy of the "best" subset.
This function applies for binary data, such as logistic regression.
}
\details{
see \code{pp_loss_binary} for additional details
about the predictive and empirical comparisons.
}
