% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{get_coefs}
\alias{get_coefs}
\title{Compute the optimal linear coefficients for any covariates}
\usage{
get_coefs(y_hat, XX, use_ols = TRUE)
}
\arguments{
\item{y_hat}{\code{n} vector of fitted values
at the given \code{XX} covariate values}

\item{XX}{\code{n x p} matrix of covariates (e.g., restricted to the
subset of \code{p} variables of interest)}

\item{use_ols}{logical; if TRUE, use ordinary least squares regression (default);
otherwise use logistic regression}
}
\value{
the \code{p} optimal regression coefficients
}
\description{
Given the fitted values from a Bayesian model and (a subset of)
covariates of interest, compute the optimal linear
coefficients. This applies for continuous (\code{use_ols = TRUE})
or binary (\code{use_ols = FALSE}) outcomes.
}
