% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{get_coefs_randint}
\alias{get_coefs_randint}
\title{Compute the optimal linear coefficients for any covariates
under a random intercept model}
\usage{
get_coefs_randint(
  post_y_pred,
  XX,
  post_sigma_e,
  post_sigma_u,
  post_y_pred_sum = NULL
)
}
\arguments{
\item{post_y_pred}{\code{S x m x n} matrix of posterior predictive draws
at the covariate values for \code{m} replicates per subject}

\item{XX}{\code{n x p} matrix of covariates (e.g., restricted to the
subset of \code{p} variables of interest)}

\item{post_sigma_e}{(\code{nsave}) draws from the posterior distribution
of the observation error SD}

\item{post_sigma_u}{(\code{nsave}) draws from the posterior distribution
of the random intercept SD}

\item{post_y_pred_sum}{(\code{nsave x n}) matrix of the posterior predictive
draws summed over the replicates within each subject (optional)}
}
\value{
the \code{p} optimal regression coefficients
}
\description{
Given the output from a random intercept model and
covariates of interest, compute the optimal linear
coefficients.
}
