% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getXY_randint}
\alias{getXY_randint}
\title{Compute the pseudo X and Y variables for LMM summarization}
\usage{
getXY_randint(
  XX,
  post_y_pred,
  post_sigma_e,
  post_sigma_u,
  post_y_pred_sum = NULL
)
}
\arguments{
\item{XX}{(\code{n x p}) matrix of covariates}

\item{post_y_pred}{(\code{nsave x m x n}) array of posterior predictive draws}

\item{post_sigma_e}{(\code{nsave}) draws from the posterior distribution
of the observation error SD}

\item{post_sigma_u}{(\code{nsave}) draws from the posterior distribution
of the random intercept SD}

\item{post_y_pred_sum}{(\code{nsave x n}) matrix of the posterior predictive
draws summed over the replicates within each subject (optional)}
}
\value{
list of the covariates and the response
}
\description{
Given output from a random intercept model, compute
the "X" and "Y" variables needed for the least squares
reparametrization.
}
