% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{initHS}
\alias{initHS}
\title{Initialize the horseshoe prior parameters}
\usage{
initHS(omega)
}
\arguments{
\item{omega}{\code{n x p} matrix of evolution errors}
}
\value{
List of relevant components: \code{sigma_wt}, the \code{n x p} matrix of standard deviations,
and the local and global precisions and parameter-expansion terms.
}
\description{
Compute the standard deviations, local and global precisions, and
parameter expansion terms for the horseshoe prior initialization.
}
