% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{lasso_path}
\alias{lasso_path}
\title{(Adaptive) lasso for Bayesian variable selection}
\usage{
lasso_path(yy_hat, XX, wts = NULL, ww_hat = NULL)
}
\arguments{
\item{yy_hat}{vector of fitted values}

\item{XX}{matrix of covariates}

\item{wts}{vector of observation weights (for weighted least squares)}

\item{ww_hat}{vector of weights for the adaptive lasso}
}
\value{
\code{inclusion_index}: the matrix of inclusion indicators (columns) for
each subset returned (rows)
}
\description{
Given output from a Bayesian model, compute an (adaptive)
lasso search path to provide linear variable selection.
This function only indicates the variables selected along the path.
}
\details{
If an intercept is in the \code{XX} matrix,
it is assumed that (i) the intercept is the first column and
(ii) the intercept should be included in each subset.
}
