% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loss_maha}
\alias{loss_maha}
\title{Compute the pseudo X and Y variables for LMM summarization}
\usage{
loss_maha(YY, y_hat, m_scale)
}
\arguments{
\item{YY}{\code{m x n} matrix of response variables}

\item{y_hat}{\code{n x 1} vector of fitted values (common across the \code{m} replicates)}

\item{m_scale}{the Mahalanobis scale factor 1/(sigma_e^2/sigma_u^2 + m)}
}
\value{
The Mahalanobis loss (scalar)
}
\description{
Given output from a random intercept model, compute
the "X" and "Y" variables needed for the least squares
reparametrization.
}
