% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{post_predict}
\alias{post_predict}
\title{Get posterior predictive draws and log-predictive density}
\usage{
post_predict(post_y_hat, post_sigma, yy = NULL)
}
\arguments{
\item{post_y_hat}{\code{nsave x n} draws of the conditional mean}

\item{post_sigma}{\code{nsave} draws of the conditional standard deviation}

\item{yy}{optional \code{n}-dimensional vector of data points; if NULL,
the lpd is not computed}
}
\value{
a list with the following elements:
\itemize{
\item \code{post_y_pred}: \code{nsave x n} posterior predictive draws
\item \code{post_lpd}: \code{nsave x n} evaluations of the log-predictive density
}
}
\description{
Given posterior samples from the conditional mean and
conditional standard deviation of a Gaussian regression model,
compute posterior predictive draws and the log-predictive
density (lpd) at the observed data points (draw-by-draw).
}
\examples{
# Simulate data:
dat = simulate_lm(n = 100, p = 10)
y = dat$y; X = dat$X

# Fit a Bayesian linear model:
fit = bayeslm::bayeslm(y ~ X[,-1], # intercept already included
              N = 1000, burnin = 500) # small sim for ex

# Compute predictive draws:
temp = post_predict(post_y_hat = tcrossprod(fit$beta, X),
                    post_sigma = fit$sigma,
                    yy = y)
names(temp) # what is returned

# Compare fitted values to the truth:
plot(dat$Ey_true,
     colMeans(temp$post_y_pred),
     xlab = 'True E(y | x)', ylab = 'Fitted')
abline(0,1)

}
