% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{pp_loss_out}
\alias{pp_loss_out}
\title{Compute the predictive squared error loss on *new* testing points}
\usage{
pp_loss_out(post_y_pred, XX, indicators, post_y_hat = NULL)
}
\arguments{
\item{post_y_pred}{\code{S x n} matrix of posterior predictive
at the given \code{XX} covariate values}

\item{XX}{\code{n x p} matrix of covariates at which to evaluate}

\item{indicators}{\code{L x p} matrix of inclusion indicators
where each row denotes a candidate subset}

\item{post_y_hat}{\code{S x n} matrix of posterior fitted values
at the given \code{XX} covariate values}
}
\value{
\code{pred_loss}: the predictive loss for each subset.
}
\description{
Use posterior predictive draws at new \code{XX} points, compute
the predictive squared error loss. The values are
computed relative to the largest subset provided, which is typically
the full set of covariates (and also the minimizer of the expected
predictive loss). These quantities are computed for a collection of
linear models that are fit to the Bayesian model output, where
each linear model features a different subset of predictors.
}
\details{
The quantity \code{post_y_hat} is the conditional expectation of the
response for each covariate value (columns) and using the parameters sampled
from the posterior (rows). For Bayesian linear regression, this term is
\code{Xbeta}. If unspecified, the algorithm will instead use \code{post_y_pred},
which is still correct but has lower Monte Carlo efficiency.
}
