% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{prescreen}
\alias{prescreen}
\title{Marginal pre-screening algorithm}
\usage{
prescreen(post_beta, num_to_keep)
}
\arguments{
\item{post_beta}{(\code{S x p}) matrix of posterior simulations
of the \code{p} regression coefficients}

\item{num_to_keep}{number of variables to return in the screening process}
}
\value{
\code{ind_keep}: the indices of the variables to keep
}
\description{
Given \code{S} draws of the \code{p} regression coefficients,
this function (i) computes (SD-standarized) posterior means,
(ii) orders the absolute values of this statistic, and
(iii) returns the indices of the largest \code{num_to_keep} values.
}
\details{
If the Bayesian model is nonlinear, \code{post_beta} may be
obtained instead by projecting each vector of posterior predictive draws
onto the covariate matrix.
}
