% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{prescreen_lasso}
\alias{prescreen_lasso}
\title{Marginal pre-screening algorithm given lasso output}
\usage{
prescreen_lasso(beta_hat_lasso, num_to_keep)
}
\arguments{
\item{beta_hat_lasso}{(\code{p x L}) matrix of estimated coefficients
from the lasso; sparsity decreases in the column index \code{L}}

\item{num_to_keep}{number of variables to return in the screening process}
}
\value{
\code{ind_keep}: the indices of the variables to keep
}
\description{
Given the estimated lasso coefficients, return the
indices of the first \code{num_to_keep} variables that enter the model
}
