% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_subsel.R
\name{proj_posterior_randint}
\alias{proj_posterior_randint}
\title{Projected predictive distribution for regression coefficients
in the random intercept model}
\usage{
proj_posterior_randint(
  post_y_pred,
  XX,
  sub_x = 1:ncol(XX),
  post_sigma_e,
  post_sigma_u,
  post_y_pred_sum = NULL
)
}
\arguments{
\item{post_y_pred}{\code{S x m x n} matrix of posterior predictive draws
at the given \code{XX} covariate values}

\item{XX}{\code{n x p} matrix of covariates}

\item{sub_x}{vector of inclusion indicators for the \code{p} covariates;
the remaining coefficients will be fixed at zero}

\item{post_sigma_e}{(\code{nsave}) draws from the posterior distribution
of the observation error SD}

\item{post_sigma_u}{(\code{nsave}) draws from the posterior distribution
of the random intercept SD}

\item{post_y_pred_sum}{(\code{nsave x n}) matrix of the posterior predictive
draws summed over the replicates within each subject (optional)}
}
\value{
\code{post_beta}: the \code{S x p} matrix of
draws from the projected predictive distribution of  the regression coefficients.
}
\description{
Given draws from the predictive distribution of the random intercept model, project these
draws onto (a  subset of) the covariates using Mahalanobis loss. This produces
many predictive draws for the regression coefficients,
which provides uncertainty quantification.
}
