% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sampleHS}
\alias{sampleHS}
\title{Sampler for horseshoe prior parameters}
\usage{
sampleHS(omega, params, sigma_e = 1)
}
\arguments{
\item{omega}{\code{n x p} matrix of errors}

\item{params}{list of parameters to update}

\item{sigma_e}{the observation error standard deviation; for (optional) scaling purposes}
}
\value{
List of relevant components in \code{params}: \code{sigma_wt}, the \code{n x p} matrix of standard deviations,
and the local and global precisions and parameter-expansion terms.
}
\description{
Compute one draw of horseshoe prior parameters (local precision, global
precision, and local and global parameter expansion terms).
}
\note{
To avoid scaling by the observation standard deviation \code{sigma_e},
simply use \code{sigma_e = 1} in the functional call.
}
