% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simulate_lm}
\alias{simulate_lm}
\title{Simulate a Gaussian linear model}
\usage{
simulate_lm(n, p, p_sig = min(5, p/2), SNR = 1)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of covariates}

\item{p_sig}{number of true nonzero coefficients (signals)}

\item{SNR}{signal-to-noise ratio}
}
\value{
a list with the following elements:
\itemize{
\item \code{y}: the response variable
\item \code{X}: the matrix of covariates
\item \code{beta_true}: the true regression coefficients (including an intercept)
\item \code{Ey_true}: the true expectation of \code{y} (\code{X\%*\%beta_true})
\item \code{sigma_true}: the true error standard deviation
}
}
\description{
Generate data from a (sparse) Gaussian linear model.
The covariates are correlated Gaussian variables. The user
may control the signal-to-noise and the number of nonzero coefficients.
}
\details{
The true regression coefficients include an intercept (-1) and
otherwise the \code{p_sig} nonzero coefficients are half equal to 1 and
half equal to -1.
}
\examples{
# Simulate data:
dat = simulate_lm(n = 100, p = 10)
names(dat) # what is returned

}
