% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simulate_lm_randint}
\alias{simulate_lm_randint}
\title{Simulate a Gaussian linear model with random intercepts}
\usage{
simulate_lm_randint(n, p, m, rho = 0.25, p_sig = min(5, p/2), SNR = 1)
}
\arguments{
\item{n}{number of subjects}

\item{p}{number of covariates}

\item{m}{number of observations per subject}

\item{rho}{intraclass correlation coefficient}

\item{p_sig}{number of true nonzero coefficients (signals)}

\item{SNR}{signal-to-noise ratio}
}
\value{
a list with the following elements:
\itemize{
\item \code{Y}: the matrix of response variables
\item \code{X}: the matrix of covariates
\item \code{beta_true}: the true regression coefficients (including an intercept)
\item \code{Ey_true}: the true expectation of \code{y} (\code{X\%*\%beta_true})
\item \code{m_scale_true}: the true Mahalanobis scale factor, 1/(sigma_e^2/sigma_u^2 + m)
}
}
\description{
Generate data from a (sparse) Gaussian linear model with random
intercepts, i.e., for repeated measurements of (longitudinal) data.
The covariates are correlated Gaussian variables. The user
may control the signal-to-noise, the number of nonzero coefficients, and
the intraclass correlation
}
\details{
The true regression coefficients include an intercept (-1) and
otherwise the \code{p_sig} nonzero coefficients are half equal to 1 and
half equal to -1.
}
\examples{
# Simulate data:
dat = simulate_lm_randint(n = 100, p = 10, m = 4)
names(dat) # what is returned

}
